chore(misc): refactor hoplite decoders to it's own module
