+++ coordinator/app/build.gradle
@@ -35,6 +35,7 @@ dependencies {
   implementation project(':jvm-libs:generic:http-rest')
   implementation project(':jvm-libs:generic:vertx-helper')
   implementation project(':jvm-libs:generic:extensions:futures')
+  implementation project(':jvm-libs:generic:extensions:hoplite')
   implementation project(':jvm-libs:generic:persistence:db')
   implementation project(':jvm-libs:linea:web3j-extensions')
   implementation project(':jvm-libs:linea:clients:eth-logs-searcher')
