+++ jvm-libs/generic/extensions/hoplite/build.gradle
@@ -1,0 +1,10 @@
+plugins {
+  id 'net.consensys.zkevm.kotlin-library-conventions'
+}
+
+description = "Hoplite configuration decoders extensions"
+
+dependencies {
+  api "com.sksamuel.hoplite:hoplite-core:${libs.versions.hoplite.get()}"
+  implementation(project(':jvm-libs:generic:extensions:kotlin'))
+}
