+++ settings.gradle
@@ -9,6 +9,7 @@ include 'jvm-libs:generic:serialization:jackson'
 include 'jvm-libs:generic:json-rpc'
 include 'jvm-libs:generic:http-rest'
 include 'jvm-libs:generic:extensions:futures'
+include 'jvm-libs:generic:extensions:hoplite'
 include 'jvm-libs:generic:extensions:kotlin'
 include 'jvm-libs:generic:extensions:tuweni'
 include 'jvm-libs:generic:logging'
